+++ src/paperglass/report/templates/__init__.py
@@ -1,0 +1,1 @@
+"""Template, stylesheet, script and shapes of the HTML report."""